+++ backend/package.json
@@ -3,9 +3,11 @@
   "name": "ekart-backend",
   "version": "1.0.0",
   "description": "backend code for the ekart project",
-  "main": "index.js",
+  "main": "app.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "node src/app.js",
+    "dev": "nodemon src/app.js"
   },
   "repository": {
     "type": "git",
@@ -28,6 +30,8 @@
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
     "express": "^4.18.2",
-    "mongoose": "^8.0.1"
+    "express-session": "^1.17.3",
+    "mongoose": "^8.0.1",
+    "nodemon": "^3.0.1"
   }
 }
